+++ devops-boot-project/devops-boot-tools/devops-boot-gradle-plugin/build.gradle.kts
@@ -11,6 +11,7 @@ dependencies {
     implementation(Libs.SpringBootGradlePlugin)
     implementation(Libs.DependencyManagement)
     implementation(Libs.KotlinSpringGradlePlugin)
+    implementation(Libs.GoogleJibPlugin)
     implementation(Libs.KtLint) {
         // ktlint 引用了1.4.31版本，和项目引入的1.4.32有冲突
         exclude(group = "org.jetbrains.kotlin", module = "kotlin-stdlib-jdk8")
