+++ devops-boot-project/devops-boot-tools/devops-boot-gradle-plugin/src/main/kotlin/com/tencent/devops/enums/AssemblyMode.kt
@@ -1,0 +1,14 @@
+package com.tencent.devops.enums
+
+enum class AssemblyMode {
+    CONSUL,
+    K8S,
+    KUBERNETES;
+
+    companion object {
+        fun ofValueOrDefault(value: String): AssemblyMode {
+            val upperCase = value.toUpperCase()
+            return values().find { it.name == upperCase } ?: CONSUL
+        }
+    }
+}
